// src/types/users-gql.types.ts
export type UsersFilterValue = 'firstName' | 'lastName' | 'age';

export interface UsersFullData {
  documentId: string;
  firstName: string;
  lastName: string;
  maidenName: string;
  age: number;
  gender: 'male' | 'female' | string;
  email: string;
  phone: string;
  username: string;
  password: string;
  birthDate: string;
  image: string;
  bloodGroup: string;
  height: number;
  weight: number;
  eyeColor: string;
  hair: {
    color: string;
    type: string;
  };
  ip: string;
  address: {
    address: string;
    city: string;
    state: string;
    stateCode: string;
    postalCode: string;
    coordinates: {
      lat: number;
      lng: number;
    };
    country: string;
  };
  macAddress: string;
  university: string;
  bank: {
    cardExpire: string;
    cardNumber: string;
    cardType: string;
    currency: string;
    iban: string;
  };
  company: {
    department: string;
    name: string;
    title: string;
    address: {
      address: string;
      city: string;
      state: string;
      stateCode: string;
      postalCode: string;
      coordinates: {
        lat: number;
        lng: number;
      };
      country: string;
    };
  };
  ein: string;
  ssn: string;
  userAgent: string;
  crypto: {
    coin: string;
    wallet: string;
    network: string;
  };
  role: string;
}

// export type UsersData = Pick<
//   UsersFullData,
//   'firstName' | 'lastName' | 'age' | 'documentId'
// >;

export interface UsersData {
  firstName: string;
  lastName: string;
  age: number;
  documentId: string;
}

export interface UserItemResponse {
  employee: UsersData;
}

export interface UsersResponse {
  employees_connection: {
    nodes: UsersFullData[];
    pageInfo: {
      page: number;
      pageSize: number;
      pageCount: number;
      total: number;
    };
  };
}

export type PersonalData = {
  id: number;
  documentId: string;
  username: string;
  email: string;
  provider: string;
  confirmed: boolean;
  blocked: boolean;
  createdAt: string;
  publishedAt: string;
  firstName: string;
  lastName: string;
};

export interface PersonalDataToken {
  jwt: string;
  user: PersonalData;
}

export interface UsersNewData {
  lastName: string;
  firstName: string;
  age: number | null;
}